simplified and PR feedback
